requirements.txt
python-telegram-bot==20.3
aiohttp==3.8.4
playwright==1.32.1
pandas==1.5.3
openpyxl==3.1.2
tqdm==4.65.0
schedule==1.2.0
python-dotenv==1.0.0
